+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - New modal - `Local Quote Statistics` with basic local quotes' stats
+- Ability to inherit listing's style in quote block (enable via settings, off by default)
 
 ### Changed
 - Danger setting got their own settings category - `Danger Zone`